refactor (*): made code more dry
Moved redundant functions to shared modules
