+++ src/lib.rs
@@ -7,3 +7,4 @@ extern crate indicatif;
 pub mod http;
 pub mod rftp;
 pub mod utils;
+pub mod bar;
