+++ src/rftp.rs
@@ -1,64 +1,13 @@
-use std::fs::File;
-use std::fs::OpenOptions;
-use std::io;
 use std::io::Read;
 use std::io::Write;
 use std::io::BufWriter;
-use std::io::ErrorKind;
-use std::fmt::Display;
 use reqwest::Url;
-use indicatif::{ProgressBar, ProgressStyle, HumanBytes};
+use indicatif::HumanBytes;
 use console::style;
-
-
 use ftp::FtpStream;
 
-static PBAR_FMT: &'static str = "{msg} {spinner:.green} {percent}% [{wide_bar:.cyan/blue}] {bytes}/{total_bytes} eta: {eta}";
-
-fn print<T: Display>(var: &T, quiet_mode: bool, is_error: bool) {
-    // print if not in quiet mode
-    if !quiet_mode {
-        if is_error {
-            eprintln!("{}", var);
-        } else {
-            println!("{}", var);
-        }
-    }
-}
-
-fn get_file_handle(fname: &str, resume_download: bool) -> io::Result<File> {
-    if resume_download {
-        match OpenOptions::new().append(true).open(fname) {
-            Ok(file) => Ok(file),
-            Err(ref error) if error.kind() == ErrorKind::NotFound => {
-                OpenOptions::new().write(true).create(true).open(fname)
-            }
-            Err(error) => Err(error),
-        }
-    } else {
-        OpenOptions::new().write(true).create(true).open(fname)
-    }
-}
-
-fn create_progress_bar(quiet_mode: bool, msg: &str, length: Option<usize>) -> ProgressBar {
-    let progbar = if quiet_mode {
-        ProgressBar::hidden()
-    } else {
-        match length {
-            Some(len) => ProgressBar::new(len as u64),
-            None => ProgressBar::new_spinner(),
-        }
-    };
-
-    progbar.set_message(msg);
-    if length.is_some() {
-        progbar.set_style(ProgressStyle::default_bar().template(PBAR_FMT).progress_chars("=> "));
-    } else {
-        progbar.set_style(ProgressStyle::default_spinner());
-    }
-
-    progbar
-}
+use utils::{get_file_handle, print};
+use bar::create_progress_bar;
 
 
 pub fn download(url: Url,
@@ -111,12 +60,20 @@ pub fn download(url: Url,
         }
     };
 
+    let bar_len = match ct_len {
+        Some(len) => {
+            let len_u64 = len as u64;
+            Some(len_u64)
+        }
+        None => None
+    };
+
 
     let chunk_size = 2048usize;
     let out_file = get_file_handle(fname, false)?;
     let mut writer = BufWriter::new(out_file);
 
-    let pbar = create_progress_bar(quiet_mode, fname, ct_len);
+    let pbar = create_progress_bar(quiet_mode, fname, bar_len);
 
     loop {
         let mut buffer = vec![0; chunk_size];
