+++ src/utils.rs
@@ -1,3 +1,7 @@
+use std::io;
+use std::fs::File;
+use std::fmt::Display;
+use std::fs::OpenOptions;
 use std::io::{Error, ErrorKind};
 use reqwest::{Url, UrlError};
 
@@ -16,3 +20,30 @@ pub fn parse_url(url: &str) -> Result<Url, UrlError> {
 pub fn gen_error(msg: String) -> Result<(), Box<::std::error::Error>> {
     Err(Box::new(Error::new(ErrorKind::Other, msg)))
 }
+
+pub fn get_file_handle(fname: &str, resume_download: bool) -> io::Result<File> {
+    if resume_download {
+        match OpenOptions::new().append(true).open(fname) {
+            Ok(file) => Ok(file),
+            Err(ref error) if error.kind() == ErrorKind::NotFound => {
+                OpenOptions::new().write(true).create(true).open(fname)
+            }
+            Err(error) => Err(error),
+        }
+    } else {
+        OpenOptions::new().write(true).create(true).open(fname)
+    }
+}
+
+pub fn print<T: Display>(var: &T, quiet_mode: bool, is_error: bool) {
+    // print if not in quiet mode
+    if !quiet_mode {
+        if is_error {
+            eprintln!("{}", var);
+        } else {
+            println!("{}", var);
+        }
+    }
+}
+
+
